+++ src/main/java/dev/ftb/mods/ftboceanmobs/entity/AbyssalWinged.java
@@ -1,7 +1,9 @@
 package dev.ftb.mods.ftboceanmobs.entity;
 
 import dev.ftb.mods.ftboceanmobs.mobai.DelayedMeleeAttackGoal;
+import dev.ftb.mods.ftboceanmobs.registry.ModSounds;
 import net.minecraft.core.BlockPos;
+import net.minecraft.sounds.SoundEvent;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.PathfinderMob;
@@ -21,6 +23,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.Nullable;
 import software.bernie.geckolib.animatable.instance.AnimatableInstanceCache;
 import software.bernie.geckolib.animation.AnimatableManager;
 import software.bernie.geckolib.animation.AnimationController;
@@ -73,6 +76,21 @@ public class AbyssalWinged extends BaseRiftMob {
     protected AABB getAttackBoundingBox() {
         // long arms...
         return super.getAttackBoundingBox().inflate(1.4);
+    }
+
+    @Override
+    protected @Nullable SoundEvent getAmbientSound() {
+        return ModSounds.ABYSSAL_WINGED_AMBIENT.get();
+    }
+
+    @Override
+    public int getAmbientSoundInterval() {
+        return 180;
+    }
+
+    @Override
+    protected SoundEvent getDeathSound() {
+        return ModSounds.ABYSSAL_WINGED_DEATH.get();
     }
 
     @Override