+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [21.1.1]
+
+### Added
+* Added many sound effects to the mobs
+
+### Changed
+* Rift Weaver boss now rapidly heals up if no players are in the arena
+  * Note: presence of creative-mode players will prevent the healing effect
+
+### Fixed
+* Fixed issue where Corrosive Craig sometimes appears to slide along ground instead of walking
+
 ## [21.1.0]
 
 ### Added